chore: create module for array handling in acirgen (#8119)
